Upgrade main libraries
