+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs 16.16.0
+nodejs 18.15.0
