+++ package.json
@@ -14,12 +14,12 @@
     "updateClusters": "npx tsc --outDir dist src/server/cluster-import && node dist/server/cluster-import.js"
   },
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.4",
-    "@mui/material": "^5.2.4",
-    "@mui/x-data-grid": "^5.2.1",
-    "@mui/x-date-pickers": "^5.0.8",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.12.0",
+    "@mui/x-data-grid": "^6.1.0",
+    "@mui/x-date-pickers": "^6.1.0",
     "@turf/bbox-polygon": "^6.5.0",
     "@turf/boolean-within": "^6.5.0",
     "@turf/center": "^6.5.0",
@@ -30,30 +30,28 @@
     "csv-stringify": "^6.0.4",
     "epsg": "^0.5.0",
     "firebase": "^9.6.1",
-    "firebase-admin": "^10.0.1",
+    "firebase-admin": "^11.6.0",
     "firebaseui": "^6.0.0",
-    "geofire-common": "^5.2.0",
-    "googleapis": "^100.0.0",
+    "geofire-common": "^6.0.0",
+    "googleapis": "^117.0.0",
     "leaflet": "^1.7.1",
     "leaflet.markercluster": "^1.5.3",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.34",
     "next": "^12.1.0",
     "proj4": "^2.7.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-firebase-hooks": "^4.0.1",
-    "react-firebaseui": "^6.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.1.1",
     "react-google-one-tap-login": "^0.1.1",
-    "react-leaflet": "^3.2.4",
-    "react-leaflet-fullscreen": "^2.0.2"
+    "react-leaflet": "^4.2.1"
   },
   "devDependencies": {
     "@types/circle-to-polygon": "^2.2.0",
     "@types/leaflet": "^1.7.8",
-    "@types/node": "16.11.13",
-    "@types/react": "17.0.37",
-    "eslint": "8.4.1",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.35",
+    "eslint": "^8.38.0",
     "eslint-config-next": "12.0.7",
     "ts-node": "^10.4.0",
     "typescript": "4.5.4"
