+++ src/components/Map/Map.tsx
@@ -1,6 +1,4 @@
 import { LayersControl, MapContainer, TileLayer, useMap } from 'react-leaflet';
-import { FullscreenControl } from 'react-leaflet-fullscreen';
-import 'react-leaflet-fullscreen/dist/styles.css';
 import { defaultPosition } from '../../hooks/constants';
 import useFirebaseLogin from '../../hooks/useFirebaseLogin';
 import Clusters from './Clusters';
@@ -76,7 +74,7 @@ export default function Map() {
           </LayersControl.Overlay>
         ))}
       </LayersControl>
-      <FullscreenControl />
+      {/* <FullscreenControl /> */}
       <UpdateMapPosition />
       <LeitungsProvider>
         <ActionButtons />
